+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- Display fraction changed instead of dollar amount.
+
 ## [0.2.0] - 2024-10-12
 
 ### Changed